Move unicode identifier tests to py3
