+++ hypothesis-python/tests/py3/test_unicode_identifiers.py
@@ -17,6 +17,7 @@
 
 from __future__ import absolute_import, division, print_function
 
+from hypothesis import given, strategies as st
 from hypothesis.internal.reflection import get_pretty_function_description, proxies
 
 
@@ -47,3 +48,13 @@ is_approx_π = lambda x: x == 3.1415  # noqa: E731
 
 def test_can_handle_unicode_identifier_in_same_line_as_lambda_def():
     assert get_pretty_function_description(is_approx_π) == "lambda x: x == 3.1415"
+
+
+def test_regression_issue_1700():
+    π = 3.1415
+
+    @given(st.floats(min_value=-π, max_value=π).filter(lambda x: abs(x) > 1e-5))
+    def test_nonzero(x):
+        assert x != 0
+
+    test_nonzero()
